# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/api/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/chat": {
      "get": {
        "tags": [
          "Chat"
        ],
        "description": "Получение списка чатов",
        "responses": {
          "200": {
            "description": "Получен список чатов",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "example": "single"
                  },
                  "pin_message": {},
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "avatar": {},
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 3
                        },
                        "datetime": {
                          "type": "string",
                          "example": "2024-04-10T02:09:40.322Z"
                        },
                        "text": {
                          "type": "string",
                          "example": "Message 2"
                        },
                        "from": {
                          "type": "number",
                          "example": 10
                        },
                        "read": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "description": "Создание чата",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Тело запроса",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "single|multi|channel"
                },
                "name": {
                  "type": "string",
                  "example": "Название чата"
                },
                "users": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Некорректные данные запроса",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "error.message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
    "/api/chat/{id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "description": "Получение чата",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "ID чата"
          }
        ],
        "responses": {
          "200": {
            "description": "Получен чат",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "type": {
                  "type": "string",
                  "example": "single"
                },
                "pin_message": {},
                "name": {
                  "type": "string",
                  "example": "test"
                },
                "avatar": {},
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 3
                      },
                      "datetime": {
                        "type": "string",
                        "example": "2024-04-10T02:09:40.322Z"
                      },
                      "text": {
                        "type": "string",
                        "example": "Message 2"
                      },
                      "from": {
                        "type": "number",
                        "example": 10
                      },
                      "read": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                },
                "users": {
                  "type": "object",
                  "properties": {
                    "1": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "datetime_last_activity": {
                          "type": "string",
                          "example": "2024-04-03T08:05:58.011Z"
                        },
                        "name": {
                          "type": "string",
                          "example": "test"
                        },
                        "login": {
                          "type": "string",
                          "example": "test"
                        },
                        "email": {},
                        "avatar": {},
                        "github_id": {}
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          },
          "403": {
            "description": "У пользователя нет доступа к чату"
          }
        }
      }
    },
...
    "/api/contact": {
      "get": {
        "tags": [
          "Contact"
        ],
        "description": "Получение списка контактов",
        "responses": {
          "200": {
            "description": "Получен список контактов",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "example": "single"
                  },
                  "pin_message": {},
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "avatar": {},
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 3
                        },
                        "datetime": {
                          "type": "string",
                          "example": "2024-04-10T02:09:40.322Z"
                        },
                        "text": {
                          "type": "string",
                          "example": "Message 2"
                        },
                        "from": {
                          "type": "number",
                          "example": 10
                        },
                        "read": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "Contact"
        ],
        "description": "Создание контакта",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Тело запроса",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "single|multi|channel"
                },
                "name": {
                  "type": "string",
                  "example": "Название чата"
                },
                "users": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Некорректные данные запроса",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "error"
                },
                "error": {
                  "type": "string",
                  "example": "error.message"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
    "/api/contact/search": {
      "get": {
        "tags": [
          "Contact"
        ],
        "description": "Получение списка контактов",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Получен список контактов",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "example": "single"
                  },
                  "pin_message": {},
                  "name": {
                    "type": "string",
                    "example": "test"
                  },
                  "avatar": {},
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 3
                        },
                        "datetime": {
                          "type": "string",
                          "example": "2024-04-10T02:09:40.322Z"
                        },
                        "text": {
                          "type": "string",
                          "example": "Message 2"
                        },
                        "from": {
                          "type": "number",
                          "example": 10
                        },
                        "read": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
...
  }
}